Added updated Lambda functions
