--- infrastructure/lambda/cleanup.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Create clean package directory
-rm -rf package/*
-rm -f lambda_function.zip
-
-echo "Clean up directories"
