+++ infrastructure/lambda/ecs/requirements.txt
@@ -1,0 +1,3 @@
+# requirements.txt
+boto3==1.34.0
+botocore==1.34.0
