+++ infrastructure/lambda/eks/cleanup.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Create clean package directory
+rm -rf package/*
+rm -f lambda_function.zip
+
+echo "Clean up directories"
