+++ infrastructure/lambda/eks/requirements.txt
@@ -1,0 +1,3 @@
+boto3==1.37.34
+kubernetes==32.0.1
+awscli==1.38.34
