--- infrastructure/lambda/requirements.txt
@@ -1,3 +1,0 @@
-boto3==1.37.33
-kubernetes==32.0.1
-awscli==1.38.34
